requirements.txt
Flask_Cors==3.0.10
Flask_MySQLdb==0.2.0
Flask==1.1.2
PyJWT==2.0.1
python-dotenv==0.15.0
